🔨 refactor algolia baking code for legibility
